refactor: Preparation for using fvttdev.
