+++ src/module/hooks/index.ts
@@ -1,0 +1,7 @@
+//NOTE! *Every* hook must be imported here, or they won't be used!
+import './preCreateActor.ts';
+import './ready.ts';
+import './renderChatMessage.ts';
+import './renderItemSheet.ts';
+import './renderSettingsConfig.ts';
+import './setup.ts';
