+++ webpack.config.ts
@@ -7,7 +7,6 @@ import MiniCssExtractPlugin from "mini-css-extract-plugin";
 
 //Only a partial type, not sure what else can be in this, and haven't looked into it.
 type FoundryConfig = { dataPath:string, systemName:string };
-const hookDir = './src/module/hooks/';
 
 function getFoundryConfig():FoundryConfig {
   const configPath = path.resolve(process.cwd(), 'foundryconfig.json');
@@ -21,7 +20,7 @@ module.exports = (env, argv) => {
   const config:Configuration = {
     context: __dirname,
     entry: {
-      main: ["./src/twodsix.ts"].concat(fs.readdirSync(hookDir).map(file => hookDir + file))
+      main: ["./src/twodsix.ts"]
     },
     mode: "development",
     module: {
